+++ src/Management/PackageDependencyCollectionOptions.ts
@@ -1,0 +1,9 @@
+import { IDependencyCollectionOptions } from "./IDependencyCollectionOptions";
+import { KeyOfType } from "./KeyOfType";
+
+/**
+ * Represents lists of dependency-names.
+ */
+export type PackageDependencyCollectionOptions = {
+    [K in KeyOfType<IDependencyCollectionOptions, Record<string, string>>]?: string[]
+};