+++ src/tests/Utilities/index.ts
@@ -1,14 +1,18 @@
+import { TestContext } from "../TestContext";
 import { JSONObjectTests } from "./JSONObject.test";
 
 /**
  * Registers tests for utilities.
+ *
+ * @param context
+ * The test-context.
  */
-export function UtilityTests(): void
+export function UtilityTests(context: TestContext): void
 {
     suite(
         "Utilities",
         () =>
         {
-            JSONObjectTests();
+            JSONObjectTests(context);
         });
 }