+++ CHANGELOG.md
@@ -6,7 +6,23 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## NPMPackageEditor [Unreleased]
 
-[Show differences](https://github.com/manuth/NPMPackageEditor/compare/v2.1.0...dev)
+[Show differences](https://github.com/manuth/NPMPackageEditor/compare/v2.2.0...dev)
+
+## NPMPackageEditor v2.2.0
+### Fixed
+  - Errors in `DependencyCollection.AllDependencies` when a dependency is occurs in multiple lists
+  - `AllDependencies` to also include `DependencyCollection.PeerDependencies`
+
+### Added
+  - A component `KeyOfType` for representing keys of properties with a predefined type
+  - A class `PackageDependencyCollection` for representing a dependency-collection which is loaded from a `Package`
+  - Support for parallel execution of drone-pipeline steps
+  - New unit-tests
+
+### Updated
+  - All dependencies
+
+[Show differences](https://github.com/manuth/NPMPackageEditor/compare/v2.1.0...v2.2.0)
 
 ## NPMPackageEditor v2.1.0
 ### Added